+++ milton/cogs/debug.py
@@ -1,5 +1,5 @@
-import aiohttp
 from discord.ext import commands
+from discord.ext.commands.context import Context
 
 from milton.bot import Milton
 from milton.utils.paginator import Paginator
@@ -7,19 +7,23 @@ from milton.utils.tools import fetch
 
 
 class DebugCog(commands.Cog, name="Debug"):
+    """"""
+
     def __init__(self, bot) -> None:
         self.bot = bot
 
-    def cog_check(self, ctx):
-        # This cog is restricted for the author of the bot.
-        # It contains error-prone functions that shouldn't be exposed
-        # to the end user
+    def cog_check(self, ctx: Context):
+        # Assure that these commands may only be executed from the owner of the
+        # bot.
         return self.bot.is_owner(ctx.author)
 
     @commands.command()
-    async def test(self, ctx):
-        """Sends a test message in order to try pagination."""
+    async def test(self, ctx: Context):
+        """Sends a test message in order to try pagination.
 
+        The contents are retrieved from a remote sever, baconipsum.com, which
+        generates meaty fillers.
+        """
         out = Paginator(title="Powered by www.baconipsum.com")
         endpt = "https://baconipsum.com/api/"
 
@@ -36,7 +40,6 @@ class DebugCog(commands.Cog, name="Debug"):
 
         for line in response.split("\n"):
             out.add_line(line)
-
         await out.paginate(ctx)