+++ milton/cogs/meta.py
@@ -3,25 +3,26 @@ import platform
 import sys
 import time
 from importlib.metadata import version
-from pathlib import Path
 
 import discord
+from discord.errors import Forbidden
 from discord.ext import commands
+from discord.ext.commands.context import Context
 
-import milton
 from milton.bot import Milton
-from milton.utils.changelog_parser import Changelog
-from milton.utils.changelog_parser import make_changelog
-from milton.utils.paginator import Paginator
+from milton.utils.errors import UserInputError
+from milton.utils.tools import get_random_line
 
 
 class MetaCog(commands.Cog, name="Meta"):
+    """Cog hosting commands related to the bot itself."""
+
     def __init__(self, bot) -> None:
         self.bot = bot
 
     @commands.command()
     @commands.cooldown(rate=1, per=10, type=commands.BucketType.user)
-    async def status(self, ctx):
+    async def status(self, ctx: Context):
         """Returns the status of the bot.
 
         Information like uptime, python version and linux version.
@@ -43,20 +44,18 @@ class MetaCog(commands.Cog, name="Meta"):
 
         embed.title = "**Milton Library Assistant - Status**"
         embed.add_field(name="Version", value=f"{me}")
-        embed.add_field(name="Python Version", value=python, inline=True)
+        embed.add_field(name="Python Version", value=python)
         embed.add_field(name="OS Information", value=f"{os}")
         embed.add_field(name="Discord.py", value=dpy)
         embed.add_field(name="Online since", value=str(since).split(".")[0])
         embed.add_field(
-            name="Watching",
-            value=f"{tot_guilds} guilds - {tot_users} users",
-            inline=True,
+            name="Watching", value=f"{tot_guilds} guilds - {tot_users} users"
         )
 
         await ctx.send(embed=embed)
 
     @commands.command()
-    async def source(self, ctx):
+    async def source(self, ctx: Context):
         """Returns the link to the bot's source code.
 
         Proudly hosted by GitHub.
@@ -69,15 +68,16 @@ class MetaCog(commands.Cog, name="Meta"):
         await ctx.send(embed=embed)
 
     @commands.command()
-    async def ping(self, ctx):
+    async def ping(self, ctx: Context):
         """Pong!
 
         Returns the discord websocket latency, kinda like the ping of the bot.
+        The real latency of the bot depends on the time it takes to process the
+        command.
         """
         embed = discord.Embed()
         embed.title = "**Pong!**"
         embed.add_field(name="Discord Websocket Latency", value=str(self.bot.latency))
-
         await ctx.send(embed=embed)
 
     @commands.group(
@@ -91,7 +91,7 @@ class MetaCog(commands.Cog, name="Meta"):
         await out.paginate(ctx)
 
     @changes_group.command()
-    async def link(self, ctx):
+    async def link(self, ctx: Context):
         """Returns the link to the bot's full changelog.
 
         For those who prefer a less interactive changelog.
@@ -103,6 +103,30 @@ class MetaCog(commands.Cog, name="Meta"):
         embed.title = "Read my full changelog on GitHub!"
         embed.url = r"https://github.com/MrHedmad/Milton/blob/master/CHANGELOG.md"
         await ctx.send(embed=embed)
+
+    @commands.command(name="inside", invoke_without_command=True)
+    @commands.guild_only()
+    async def subscribe(self, ctx: Context):
+        """Subscribe to announcements and other things."""
+        if (role := ctx.guild.get_role(777612764222717992)) :
+            if role not in ctx.author.roles:
+                try:
+                    await ctx.author.add_roles(role)
+                except Forbidden:
+                    raise UserInputError("I do not have powers here... :(")
+                embed = discord.Embed(title="You are now inside!")
+                embed.set_image(url=get_random_line("./milton/resources/insiders.txt"))
+                await ctx.send(embed=embed)
+            else:
+                try:
+                    await ctx.author.remove_roles(role)
+                except Forbidden:
+                    raise UserInputError("I do not have powers here... :(")
+                embed = discord.Embed(title="You are no longer inside.")
+                embed.set_image(
+                    url=get_random_line("./milton/resources/noinsiders.txt")
+                )
+                await ctx.send(embed=embed)
 
 
 def setup(bot: Milton):