+++ milton/utils/checks.py
@@ -1,11 +1,17 @@
 """Implements custom checks to be used as decorators"""
+from typing import Callable
+
 from discord.ext.commands import check
 
 from milton.config import CONFIG
 
 
-def in_guild(id):
-    """A Check that sees if the guild is the same as specified."""
+def in_guild(id) -> Callable:
+    """A Check that sees if the guild is the same as specified.
+
+    Args:
+        id: The id of the guild to check for
+    """
 
     def predicate(ctx):
         return ctx.guild.id == id
@@ -13,8 +19,12 @@ def in_guild(id):
     return check(predicate)
 
 
-def in_channel(id):
-    """A Check that sees if the channel is the same as specified."""
+def in_channel(id) -> Callable:
+    """A Check that sees if the channel is the same as specified.
+
+    Args:
+        id: The id of the channel to check for
+    """
 
     def predicate(ctx):
         return ctx.channel.id == id
@@ -22,8 +32,13 @@ def in_channel(id):
     return check(predicate)
 
 
-def in_test_guild():
-    """A Check that sees if the channel is the same as specified."""
+def in_test_guild() -> Callable:
+    """A Check that sees if the message was sent in the test guild.
+
+    The test guild is the one specified in the configuration.
+
+    TODO: What if no guild is specified? This just crashes
+    """
 
     def predicate(ctx):
         return ctx.guild.id == CONFIG.bot.test_server_id